pyproject.toml
[project]
name = "ifd-cpb-python-ai"
version = "0.1.1"
requires-python = ">=3.13"
dependencies = [
    "aiofiles",
    "asyncpg",
    "atlassian-python-api",
    "beautifulsoup4",
    "fastapi",
    "google-cloud-pubsub",
    "google-cloud-secret-manager",
    "google-cloud-storage",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-openai",
    "langchain-postgres",
    "langchain-text-splitters",
    "langgraph",
    "langgraph-checkpoint-postgres",
    "markdown",
    "marshmallow",
    "numpy",
    "openai",
    "pandas",
    "psycopg",
    "psycopg-binary",
    "psycopg2-binary",
    "pydantic",
    "pydantic-core",
    "python-dotenv",
    "python-multipart",
    "rank-bm25",
    "requests",
    "sqlalchemy",
    "sqlalchemy-orm",
    "sqlalchemy-utils",
    "tabulate",
    "tavily-python>=0.7.9",
    "teams-ai",
    "unstructured[docx]",
    "uvicorn",
    "websockets",
]

[dependency-groups]
dev = [
    "coverage>=7.0.0",
    "factory-boy>=3.3.3",
    "ipykernel>=6.30.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0", 
    "pytest-cov>=6.2.1",
    "pytest-md-report>=0.7.0",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
    ".",
]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",
    
    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",
    
    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",
    
    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
]

ignore_errors = true
show_missing = true

[tool.coverage.html]
directory = "htmlcov"
